Fixing VOD (Shows & Movies) errors caused by IPTV incomplete data
The IPTV provider returned data seems to be partially complete, or missing pieces that this proxy is expecting.
The unmarshalling processes was failing mainly when the 'info', 'video' & 'audio' JSON data portions were empty or missing. These fields now have their own unmarshalling methods and if an error is encountered during this process, then it attempts to unmarshall the fields individually in the hope of returning as much data as possible.
If additional errors occur these should now be logged.